Fix : 패키징 명칭 변경 및 경로 일부 변경
